chore: complete readme to get to the point tests pass
- added a `build` commands to composer which compiles horizon & telescope
- update readme including openssl workaround (based on node version in `Dockerimage`)

Note: `yarn` will not run with `--frozen-lockfile` with the lockfile currently in git.
Yarn is run during deployment so I assume an updated lockfile should be pushed into vc
but have not included it in this commit from an abundance of caution.

Note: horizon / telescope included in build due to their presence in the test ci workflow.
However, they don't seem to be updated as part of the deployment workflow.
Their published files are under version control but these changes have again been excluded from this commit.