TECH-245 Return the gatwick requirements page
Took a simple approach of not using the existing requirements systems as this seems to be defunct post gcap.

Instead just looked for 50 hours on gnd/del in the UK from the atc history data and checked roster status.

Roster functionality untested as getting a test account onto the roster was beyond my ken, but tests cover hour check at least.