refactor: Large refactoring to WaitingListAccount, converting it from a pivot table to a full status entity
This is quite a large commit that moves the AccountWaitingList from a `Pivot` to an `Entity` and changes some related functionality.

Doing so allows further work on waiting lists in the future and enables some cleanups to waiting list code.

As a result of this it is somewhat harder to jump up and down (e.g access WaitingLists from Accounts) but all functionality should remain the same. I tried to has laravels Has...Through type relationships but they don't work for this tree unfortunately