Contributing Guides (#4087)
* Added contributing guides

Added Guides to provide help to those interested in working on the UK core application.

- Updated Contributing
- Added First time contributors guide
- Added Issue Template
- Added Pull Request Template
- Added Style guide from Sector File
- Added the contents of readme to testing
- Reduced the text in readme

* Update .github/CONTRIBUTING.md

Co-authored-by: CJA <git@callumaxon.uk>

* Update .github/CONTRIBUTING.md

Co-authored-by: CJA <git@callumaxon.uk>

* Update .github/CONTRIBUTING.md

Co-authored-by: CJA <git@callumaxon.uk>

* Update .github/CONTRIBUTING.md

Co-authored-by: CJA <git@callumaxon.uk>

* Update .github/CONTRIBUTING.md

Co-authored-by: CJA <git@callumaxon.uk>

* Update .github/testing.md

Co-authored-by: CJA <git@callumaxon.uk>

* Update testing.md

Omitted larvel version from URL as its not necessary

* Update First time contributors guide.md

Replaced a majority of the images

Still required to get a few more changes

* Update testing.md

Due to the ever-evolving nature of both PHP and Nodejs I have changed to link to their respective JSON config files.

* Update First time contributors guide.md

* Modified package json

Included engine version for package.json

as linked within the Documentation

---------

Co-authored-by: CJA <git@callumaxon.uk>